+++ src/main/java/dave/commands/DeleteTaskCommand.java
@@ -1,12 +1,12 @@
-package commands;
+package dave.commands;
 
 import dave.Storage;
 import dave.TaskList;
 import dave.Ui;
 // import exceptions.ChatbotException;
 // import exceptions.InvalidInputException;
-import exceptions.UnableToFindTaskException;
-import tasks.*;
+import dave.exceptions.UnableToFindTaskException;
+import dave.tasks.*;
 
 public class DeleteTaskCommand extends Command {
     private int taskNumber;
