+++ src/main/java/dave/commands/ExitCommand.java
@@ -1,4 +1,4 @@
-package commands;
+package dave.commands;
 
 import dave.Storage;
 import dave.TaskList;
