+++ src/main/java/dave/exceptions/ChatbotException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package dave.exceptions;
 public class ChatbotException extends Exception {
     public ChatbotException(String msg) {
         super(msg);
