+++ src/main/java/dave/exceptions/EmptyTaskException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package dave.exceptions;
 public class EmptyTaskException extends ChatbotException {
     public EmptyTaskException(String msg) {
         super(msg);
