+++ src/main/java/dave/exceptions/UnableToFindTaskException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package dave.exceptions;
 
 import dave.TaskList;
