+++ src/main/java/dave/tasks/Event.java
@@ -1,4 +1,4 @@
-package tasks;
+package dave.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
